Viscacha 0.8 RC3

+++ classes/cron/jobs/deletesearch.php
@@ -6,7 +6,7 @@ global $filesystem;
 $dir = 'cache/search/';
 $d = dir($dir);
 while (false !== ($entry = $d->read())) {
-	if (!is_dir($dir.$entry) && file_exists($dir.$entry) && fileatime($dir.$entry) < time()-60*60*24*7) {
+	if ($entry != '.htaccess' && $entry != 'index.htm' && !is_dir($dir.$entry) && file_exists($dir.$entry) && fileatime($dir.$entry) < time()-60*60*24*7) {
 		$filesystem->unlink($dir.$entry);
 	}
 }
